+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## Upcoming
 
-## Font changes
+### Font changes
 
 - Default primary font is now Stena Sans, with Open Sans fallback.
 
-## Design changes
+### Design changes
 
 - Increased contrast in: `Label`, `ValueTable` and input placeholders.
 - Buttons now have a `width` prop which is responsive according to styled-system.
@@ -14,6 +14,10 @@
 ## Fixes
 
 Remove letter-spacing override done by setting text size.
+
+### Internal
+
+- `Space` now uses emotion for styling instead of inline styles.
 
 ## 8.0.2